+++ include/souls_formats/sf_common.h
@@ -19,9 +19,9 @@ extern "C" {
  *===========================================================================*/
 
 #define SF_VERSION_MAJOR 0
-#define SF_VERSION_MINOR 6
+#define SF_VERSION_MINOR 8
 #define SF_VERSION_PATCH 0
-#define SF_VERSION_STRING "0.6.0"
+#define SF_VERSION_STRING "0.8.0"
 
 /*===========================================================================
  * ABI export macro